#26 Expose datetime / struct_time conversion functions
